+++ cli/integrationtest/controlplane_test.go
@@ -389,11 +389,13 @@ timeoutSeconds: 600`, getTestConnectivityImage(t))
 	runSpecPath := filepath.Join(tempDir, "runspec.yaml")
 	require.NoError(os.WriteFile(runSpecPath, []byte(runSpec), 0644))
 
-	execStdOut := NewTygerCmdBuilder("run", "exec", "--file", runSpecPath, "--log-level", "trace").
+	execStdOut, execStdErr, err := NewTygerCmdBuilder("run", "exec", "--file", runSpecPath, "--logs", "--log-level", "trace").
 		Stdin("0123").
-		RunSucceeds(t)
+		Run()
 
+	require.NoError(err)
 	require.Equal("1234", execStdOut)
+	require.NotContains(strings.ToLower(execStdErr), "timed out waiting for logs")
 }
 
 func TestEndToEndExecWithSocketsAndEphemeralBuffers(t *testing.T) {
